Fase 3.4: threading do idMunicipio — 11 de 36 dominios do Betim
Cada funcao de lib/betim migrada recebe idMunicipio como PRIMEIRO
parametro obrigatorio, e as paginas o resolvem com cidadeDaRota(params) —
que traduz o SLUG da rota (`betim`) para o id_municipio (`3106705`) num
lugar so, em vez de repetir o boilerplate em 45 paginas.

Migrados: caixa, obras, social, comercios, indicadores, seguranca,
verbas, notaTransparencia, educacao, postos, servidores.
ID_MUNICIPIO_DEFAULT: 124 -> 102 ocorrencias.

TRES ARMADILHAS DE TIPO, todas achadas por verificacao e nao por leitura:

1. Colunas `numeric` voltam como STRING no driver e como number no
   PostgREST. Criei lib/db/num.ts para converter, e um AUDITOR que
   confere cada num(tabela.coluna) contra o tipo real do schema. Ele pegou
   um erro meu: eu havia convertido `indicadores.valor`, que e TEXT, nao
   numeric — converter teria corrompido o dado. O levantamento manual que
   me levou a isso misturava colunas de tabelas vizinhas.

2. `count(*) over ()` e bigint, e bigint tambem volta como STRING. O total
   de servidores vinha "9803" em vez de 9803. Passa despercebido em
   `total / 50` (o JS coage) mas `total + 1` concatenaria. Corrigido com
   cast ::int aqui e no Congresso.

3. O compilador virou a rede de seguranca prometida: ao trocar a
   assinatura, ele apontou exatamente as 13 paginas que ainda nao passavam
   a cidade. Nenhuma passou despercebida.

Verificado: scripts/paridade-betim.mts com 13 comparacoes contra o
Supabase, todas iguais, incluindo somas com centavos (obras 448765823.33,
verbas 89173.18). Build verde, 68 paginas.

+++ apps/web/lib/db/queries/congresso.ts
@@ -340,7 +340,7 @@ export async function paginaDeProposicoes(filtros: FiltrosProposicoes = {}) {
     .select({
       proposicao: proposicoesInCongresso,
       analise: analisesInCongresso,
-      total: sql<number>`count(*) over ()`,
+      total: sql<number>`(count(*) over ())::int`,
     })
     .from(proposicoesInCongresso)
     .leftJoin(
